Add A1_T5 solution

+++ A1_T5.py
@@ -4,11 +4,11 @@ print("Calculate the area of a wall.")
 
 # Ask for width
 Feed = input("Enter the width in meters: ")
-Width = int(Feed)   # convert to integer
+Width = float(Feed)   # convert to integer
 
 # Ask for height
 Feed = input("Enter the height in meters: ")
-Height = int(Feed)  # convert to integer
+Height = float(Feed)  # convert to integer
 
 # Show width and height
 print(f"Width is {Width} m and height is {Height} m.")
